separate file for vector functions of ieee
